+++ src/main/java/ua/khnu/service/GroupService.java
@@ -5,9 +5,9 @@ import ua.khnu.entity.Group;
 import java.util.List;
 
 public interface GroupService {
-    void createNewGroup(int userId, String message);
+    void createNewGroup(long userId, String message);
 
     List<Group> getAllGroups();
 
-    List<Group> getUserGroups(int userId);
+    List<Group> getUserGroups(long userId);
 }
